refactor: shell dir
