+++ src/kernel/shell/cmd_hello.cpp
@@ -1,5 +1,5 @@
 #include "board/mini-uart.hpp"
-#include "cmd.hpp"
+#include "shell/cmd.hpp"
 
 int cmd_hello(int /* argc */, char* /* argv */[]) {
   mini_uart_puts("Hello World!\n");
