+++ src/kernel/shell/cmd_ls.cpp
@@ -1,6 +1,6 @@
 #include "board/mini-uart.hpp"
-#include "cmd.hpp"
 #include "fs/initramfs.hpp"
+#include "shell/cmd.hpp"
 
 int cmd_ls(int /* argc */, char* /* argv */[]) {
   for (auto it : initramfs) {
