+++ src/kernel/shell/cmd_run.cpp
@@ -1,7 +1,7 @@
 #include "board/mini-uart.hpp"
-#include "cmd.hpp"
 #include "exec.hpp"
 #include "fs/initramfs.hpp"
+#include "shell/cmd.hpp"
 
 extern char __user_text[];
 extern char __user_stack[];
